Fixed $this error - added class structure & example
