+++ app/Http/Controllers/HomeController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Restaurant;
+use \stdClass;
 
 class HomeController extends Controller
 {
     public function index()
     {
-        
+
+
         $restaurants = Restaurant::all();
 
         return view('home', [
