+++ app/Http/Controllers/RestaurantController.php
@@ -4,35 +4,20 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Restaurant;
+use App\classes\Backend;
 
 
 class RestaurantController extends Controller
 {
 
-    public function generateSlug($slug){
-      //remove any character that isnt a space hyphen letter number
-      //remove spaces and duplicate hyphens
-      //trim left and right from any extra hyphens
-
-      $lettersNumbersSpacesHyphens = '/[^-/sa-zA-Z0-9]/';
-      $spacesDuplicateHyphens = '/[-/s]+/';
-
-      $slug = preg_replace($lettersNumbersSpacesHyphens, '', $slug);
-      $slug = preg_replace($spacesDuplicateHyphens, '-', $slug);
-      $slug = trim($slug, '-');
-
-      return $slug;
-
-    }
-
     public function create(Request $request)
     {
 
       $restaurant = new Restaurant();
+      $_backend = new Backend();
       $restaurant->name = $request->name;
       $restaurant->description = $request->description;
-      $restaurant->slug = generateSlug($request->name);
-
+      $restaurant->slug = $_backend->generateSlug($request->name);
 
       $restaurant->save();
